SAWScript.Crucible.{LLVM,JVM,MIR}: Move extension-related data types to *.Setup.Value modules
This is purely a refactor. This will become useful in a subsequent commit in
order to avoid import cycles.
