+++ src/SAWScript/Crucible/JVM/MethodSpecIR.hs
@@ -61,8 +61,6 @@ import qualified Prettyprinter as PPL
 -- what4
 import           What4.ProgramLoc (ProgramLoc)
 
-import qualified Lang.Crucible.FunctionHandle as Crucible (HandleAllocator)
-
 -- crucible-jvm
 import qualified Lang.Crucible.JVM as CJ
 import qualified Lang.JVM.Codebase as CB
@@ -70,47 +68,13 @@ import qualified Lang.JVM.Codebase as CB
 -- jvm-parser
 import qualified Language.JVM.Parser as J
 
--- cryptol-saw-core
-import           Verifier.SAW.TypedTerm (TypedTerm)
-
 import           SAWScript.Crucible.Common
 import qualified SAWScript.Crucible.Common.MethodSpec as MS
 import qualified SAWScript.Crucible.Common.Setup.Type as Setup
-
---------------------------------------------------------------------------------
--- ** Language features
-
-type instance MS.HasSetupNull CJ.JVM = 'True
-type instance MS.HasSetupGlobal CJ.JVM = 'False
-type instance MS.HasSetupStruct CJ.JVM = 'False
-type instance MS.HasSetupArray CJ.JVM = 'False
-type instance MS.HasSetupElem CJ.JVM = 'False
-type instance MS.HasSetupField CJ.JVM = 'False
-type instance MS.HasSetupCast CJ.JVM = 'False
-type instance MS.HasSetupUnion CJ.JVM = 'False
-type instance MS.HasSetupGlobalInitializer CJ.JVM = 'False
-
-type instance MS.HasGhostState CJ.JVM = 'False
-
-type JIdent = String -- FIXME(huffman): what to put here?
-
-type instance MS.TypeName CJ.JVM = JIdent
-
-type instance MS.ExtType CJ.JVM = J.Type
-type instance MS.CastType CJ.JVM = ()
-type instance MS.ResolvedState CJ.JVM = ()
+import           SAWScript.Crucible.JVM.Setup.Value
 
 --------------------------------------------------------------------------------
 -- *** JVMMethodId
-
-data JVMMethodId =
-  JVMMethodId
-    { _jvmMethodKey :: J.MethodKey
-    , _jvmClassName  :: J.ClassName
-    }
-  deriving (Eq, Ord, Show)
-
-makeLenses ''JVMMethodId
 
 jvmMethodName :: Getter JVMMethodId String
 jvmMethodName = jvmMethodKey . to J.methodKeyName
@@ -121,20 +85,8 @@ csMethodKey = MS.csMethod . jvmMethodKey
 csMethodName :: Getter (MS.CrucibleMethodSpecIR CJ.JVM) String
 csMethodName = MS.csMethod . jvmMethodName
 
--- TODO: avoid intermediate 'String' values
-instance PPL.Pretty JVMMethodId where
-  pretty (JVMMethodId methKey className) =
-    PPL.pretty (concat [J.unClassName className ,".", J.methodKeyName methKey])
-
-type instance MS.MethodId CJ.JVM = JVMMethodId
-
 --------------------------------------------------------------------------------
 -- *** Allocation
-
-data Allocation
-  = AllocObject J.ClassName
-  | AllocArray Int J.Type
-  deriving (Eq, Ord, Show)
 
 allocationType :: Allocation -> J.Type
 allocationType alloc =
@@ -142,20 +94,8 @@ allocationType alloc =
     AllocObject cname -> J.ClassType cname
     AllocArray _len ty -> J.ArrayType ty
 
-
--- TODO: We should probably use a more structured datatype (record), like in LLVM
-type instance MS.AllocSpec CJ.JVM = (MS.ConditionMetadata, Allocation)
-
 --------------------------------------------------------------------------------
 -- *** PointsTo
-
-type instance MS.PointsTo CJ.JVM = JVMPointsTo
-
-data JVMPointsTo
-  = JVMPointsToField MS.ConditionMetadata MS.AllocIndex J.FieldId (Maybe (MS.SetupValue CJ.JVM))
-  | JVMPointsToStatic MS.ConditionMetadata J.FieldId (Maybe (MS.SetupValue CJ.JVM))
-  | JVMPointsToElem MS.ConditionMetadata MS.AllocIndex Int (Maybe (MS.SetupValue CJ.JVM))
-  | JVMPointsToArray MS.ConditionMetadata MS.AllocIndex (Maybe TypedTerm)
 
 overlapPointsTo :: JVMPointsTo -> JVMPointsTo -> Bool
 overlapPointsTo =
@@ -206,22 +146,6 @@ instance PPL.Pretty JVMPointsTo where
 
 --------------------------------------------------------------------------------
 -- *** JVMCrucibleContext
-
-type instance MS.Codebase CJ.JVM = CB.Codebase
-
-data JVMCrucibleContext =
-  JVMCrucibleContext
-  { _jccJVMClass       :: J.Class
-  , _jccCodebase       :: CB.Codebase
-  , _jccJVMContext     :: CJ.JVMContext
-  , _jccBackend        :: SomeOnlineBackend
-  , _jccHandleAllocator :: Crucible.HandleAllocator
-  }
-
-makeLenses ''JVMCrucibleContext
-
-type instance MS.CrucibleContext CJ.JVM = JVMCrucibleContext
-
 
 jccWithBackend ::
   JVMCrucibleContext ->
