+++ src/SAWScript/Crucible/MIR/MethodSpecIR.hs
@@ -41,15 +41,9 @@ module SAWScript.Crucible.MIR.MethodSpecIR
   , initialCrucibleSetupState
   ) where
 
-import Control.Lens (Getter, (^.), makeLenses, to)
-import Data.Parameterized.Classes
-import Data.Parameterized.Some
-import Data.Text (Text)
+import Control.Lens (Getter, (^.), to)
 import qualified Prettyprinter as PP
 
-import Lang.Crucible.FunctionHandle (HandleAllocator)
-import Lang.Crucible.Types
-import Mir.DefId
 import Mir.Generator
 import Mir.Intrinsics
 import qualified Mir.Mir as M
@@ -57,40 +51,8 @@ import What4.ProgramLoc (ProgramLoc)
 
 import           SAWScript.Crucible.Common
 import qualified SAWScript.Crucible.Common.MethodSpec as MS
-import qualified SAWScript.Crucible.Common.Override as MS
 import qualified SAWScript.Crucible.Common.Setup.Type as Setup
-
-type instance MS.HasSetupNull MIR = 'False
-type instance MS.HasSetupGlobal MIR = 'False
-type instance MS.HasSetupStruct MIR = 'True
-type instance MS.HasSetupArray MIR = 'True
-type instance MS.HasSetupElem MIR = 'True
-type instance MS.HasSetupField MIR = 'True
-type instance MS.HasSetupCast MIR = 'False
-type instance MS.HasSetupUnion MIR = 'False
-type instance MS.HasSetupGlobalInitializer MIR = 'False
-
-type instance MS.HasGhostState MIR = 'False
-
-type instance MS.TypeName MIR = Text
-type instance MS.ExtType MIR = M.Ty
-
-type instance MS.MethodId MIR = DefId
-type instance MS.AllocSpec MIR = Some MirAllocSpec
-type instance MS.PointsTo MIR = MirPointsTo
-type instance MS.ResolvedState MIR = ()
-type instance MS.CastType MIR = ()
-
-type instance MS.Codebase MIR = CollectionState
-
-data MIRCrucibleContext =
-  MIRCrucibleContext
-  { _mccRustModule      :: RustModule
-  , _mccBackend         :: SomeOnlineBackend
-  , _mccHandleAllocator :: HandleAllocator
-  }
-
-type instance MS.CrucibleContext MIR = MIRCrucibleContext
+import           SAWScript.Crucible.MIR.Setup.Value
 
 mccWithBackend ::
   MIRCrucibleContext ->
@@ -102,43 +64,11 @@ mccWithBackend cc k =
 mccSym :: Getter MIRCrucibleContext Sym
 mccSym = to (\mcc -> mccWithBackend mcc backendGetSym)
 
-type instance MS.Pointer' MIR sym = Some (MirPointer sym)
-
--- | Unlike @LLVMPointsTo@ and @JVMPointsTo@, 'MirPointsTo' contains a /list/ of
--- 'MS.SetupValue's on the right-hand side. This is due to how slices are
--- represented in @crucible-mir-comp@, which stores the list of values
--- referenced by the slice. The @mir_points_to@ command, on the other hand,
--- always creates 'MirPointsTo' values with exactly one value in the list (see
--- the @firstPointsToReferent@ function in "SAWScript.Crucible.MIR.Override").
-data MirPointsTo = MirPointsTo MS.ConditionMetadata MS.AllocIndex [MS.SetupValue MIR]
-    deriving (Show)
-
 instance PP.Pretty MirPointsTo where
     pretty (MirPointsTo _md alloc sv) = PP.parens $
         PP.pretty (show alloc) PP.<+> "->" PP.<+> PP.list (map MS.ppSetupValue sv)
 
-data MirAllocSpec tp = MirAllocSpec
-    { _maType :: TypeRepr tp
-    , _maMutbl :: M.Mutability
-    , _maMirType :: M.Ty
-    , _maLen :: Int
-    }
-  deriving (Show)
-
-instance ShowF MirAllocSpec where
-
-data MirPointer sym tp = MirPointer
-    { _mpType :: TypeRepr tp
-    , _mpMutbl :: M.Mutability
-    , _mpMirType :: M.Ty
-    , _mpRef :: MirReferenceMux sym tp
-    }
-
 type MIRMethodSpec = MS.CrucibleMethodSpecIR MIR
-
-makeLenses ''MIRCrucibleContext
-makeLenses ''MirAllocSpec
-makeLenses ''MirPointer
 
 initialDefCrucibleMethodSpecIR ::
   CollectionState ->
